% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{convert}
\alias{convert}
\title{Convert object of on class to another}
\usage{
convert(object, class_out = NULL, ..., converter = NULL)
}
\arguments{
\item{object}{object to be converted}

\item{class_out}{class to be converted to}

\item{...}{additional arguments to be passed to the converter}

\item{converter}{optional override converter to bypass default one}
}
\description{
Convert object of on class to another
}
