% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\docType{class}
\name{converter}
\alias{converter}
\title{Class to interact with default converters}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
converter
}
\arguments{
\item{class_in}{class to be converted}

\item{class_out}{class to be converted to}

\item{f}{function that acts as converter}
}
\description{
Class to interact with default converters
}
\section{Methods}{

\code{$get(class_in, class_out)} Get default converter for a given conversion from \code{class_in} to \code{class_out}
\code{$set(class_in, class_out, f)} Set the default converter
}

\keyword{datasets}
