% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_dataset.R
\name{from_schema.arrow_stream_dataset}
\alias{from_schema.arrow_stream_dataset}
\title{Create an Arrow Dataset from an input stream, inferring output types and
shapes from the given Arrow schema.}
\usage{
\method{from_schema}{arrow_stream_dataset}(object, schema, columns = NULL, host = NULL, filenames = NULL, ...)
}
\arguments{
\item{object}{An \R object.}

\item{schema}{Arrow schema defining the record batch data in the stream.}

\item{columns}{A list of column indices to be used in the Dataset.}

\item{host}{A \code{tf.string} tensor or string defining the input stream.
For a socket client, use "<HOST_IP>:<PORT>", for stdin use "STDIN".}

\item{filenames}{Not used.}

\item{...}{Optional arguments passed on to implementing methods.}
}
\description{
Create an Arrow Dataset from an input stream, inferring output types and
shapes from the given Arrow schema.
}
