% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_dataset.R
\name{from_schema}
\alias{from_schema}
\title{Create an Arrow Dataset from the given Arrow schema.}
\usage{
from_schema(object, ...)
}
\arguments{
\item{object}{An \R object.}

\item{...}{Optional arguments passed on to implementing methods.}
}
\description{
Infer output types and shapes from the given Arrow schema and create an Arrow
Dataset.
}
