% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_dataset.R
\name{from_schema.tensorflow_io.arrow.python.ops.arrow_dataset_ops.ArrowFeatherDataset}
\alias{from_schema.tensorflow_io.arrow.python.ops.arrow_dataset_ops.ArrowFeatherDataset}
\title{Create an Arrow Dataset for reading record batches from Arrow feather files,
inferring output types and shapes from the given Arrow schema.}
\usage{

  \method{from_schema}{tensorflow_io.arrow.python.ops.arrow_dataset_ops.ArrowFeatherDataset}(object,
  filenames, schema, columns = NULL)
}
\arguments{
\item{filenames}{A \code{tf.string} tensor, list or scalar containing files in
Arrow Feather format.}

\item{schema}{Arrow schema defining the record batch data in the stream.}

\item{columns}{A list of column indices to be used in the Dataset.}
}
\description{
Create an Arrow Dataset for reading record batches from Arrow feather files,
inferring output types and shapes from the given Arrow schema.
}
