% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrow_dataset.R
\name{from_schema.tensorflow_io.arrow.python.ops.arrow_dataset_ops.ArrowStreamDataset}
\alias{from_schema.tensorflow_io.arrow.python.ops.arrow_dataset_ops.ArrowStreamDataset}
\title{Create an Arrow Dataset from an input stream, inferring output types and
shapes from the given Arrow schema.}
\usage{

  \method{from_schema}{tensorflow_io.arrow.python.ops.arrow_dataset_ops.ArrowStreamDataset}(object,
  host, schema, columns = NULL)
}
\arguments{
\item{host}{A \code{tf.string} tensor or string defining the input stream.
For a socket client, use "<HOST_IP>:<PORT>", for stdin use "STDIN".}

\item{schema}{Arrow schema defining the record batch data in the stream.}

\item{columns}{A list of column indices to be used in the Dataset.}
}
\description{
Create an Arrow Dataset from an input stream, inferring output types and
shapes from the given Arrow schema.
}
