% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lut.R
\docType{class}
\name{io_lut}
\alias{io_lut}
\title{Input/output look up table}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
io_lut
}
\description{
Input/output look up table
}
\examples{



}
\keyword{datasets}
