% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supported.R
\name{io_supported}
\alias{io_supported}
\title{Determine input-output support for data or file type}
\usage{
io_supported(type)
}
\arguments{
\item{type}{data or file type}
}
\value{
a \code{data.frame} with logical entries;
        \code{TRUE} if type is supported, \code{FALSE} otherwise
}
\description{
This function returns whether a type is supported by
\code{\link{qread}} or \code{\link{qwrite}}.
}
\examples{
io_supported("rds")

}

