% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmdb_dataset.R
\name{lmdb_dataset}
\alias{lmdb_dataset}
\title{Create a \code{LMDBDataset}.}
\usage{
lmdb_dataset(filenames)
}
\arguments{
\item{filenames}{A \code{tf.string} tensor containing one or more filenames.}
}
\description{
This function allows a user to read data from a LMDB
file. A lmdb file consists of (key value) pairs sequentially.
}
\examples{
\dontrun{
dataset <- sequence_file_dataset("testdata/data.mdb") \%>\%
  dataset_repeat(1)

sess <- tf$Session()
iterator <- make_iterator_one_shot(dataset)
next_batch <- iterator_get_next(iterator)

until_out_of_range({
  batch <- sess$run(next_batch)
  print(batch)
})
}

}
