% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnist_dataset.R
\name{mnist_label_dataset}
\alias{mnist_label_dataset}
\title{Creates a \code{MNISTLabelDataset}.}
\usage{
mnist_label_dataset(filenames, compression_type = NULL)
}
\arguments{
\item{filenames}{A \code{tf.string} tensor containing one or more filenames.}

\item{compression_type}{A \code{tf.string} scalar evaluating to one
of \code{""} (no compression), \code{"ZLIB"}, or \code{"GZIP"}.}
}
\description{
This creates a dataset for MNIST labels.
}
