% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquet_dataset.R
\name{parquet_dataset}
\alias{parquet_dataset}
\title{Create a \code{ParquetDataset}.}
\usage{
parquet_dataset(filenames, columns, output_types)
}
\arguments{
\item{filenames}{A 0-D or 1-D \code{tf.string} tensor containing one or more
filenames.}

\item{columns}{A 0-D or 1-D \code{tf.int32} tensor containing the columns to
extract.}

\item{output_types}{A tuple of \code{tf.DType} objects representing the types of
the columns returned.}
}
\description{
This allows a user to read data from a parquet file.
}
