% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubsub_dataset.R
\name{pubsub_dataset}
\alias{pubsub_dataset}
\title{Creates a \code{PubSubDataset}.}
\usage{
pubsub_dataset(subscriptions, server = NULL, eof = FALSE, timeout = 1000)
}
\arguments{
\item{subscriptions}{A \code{tf.string} tensor containing one or more
subscriptions.}

\item{server}{The pubsub server.}

\item{eof}{If True, the pubsub reader will stop on EOF.}

\item{timeout}{The timeout value for the PubSub to wait (in millisecond).}
}
\description{
This creates a dataset for consuming PubSub messages.
}
