% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read}
\alias{read}
\title{Read stuff into R}
\usage{
read(file, extension = tools::file_ext(file), ..., reader = NULL)
}
\arguments{
\item{file}{the name of the file which the data are to be read from}

\item{extension}{the file extension (e.g. '.csv') used to find the default reader}

\item{...}{additional arguments to be passed to the reader}

\item{reader}{optional override reader to bypass default one}
}
\description{
Read stuff into R
}
\examples{
tf <- tempfile()
writeLines(test1, tf)
df <- read(tf, fill = TRUE)
}
