% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R, R/write.R
\docType{class}
\name{reader}
\alias{reader}
\alias{writer}
\title{Class to interact with default readers}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
reader

writer
}
\arguments{
\item{extension}{file extension (class character)}

\item{f}{function that acts as reader}

\item{class_in}{the class of data to be written to file}

\item{extension}{file extension (class character)}

\item{f}{function that acts as writer}
}
\description{
Class to interact with default readers

Class to interact with default writers
}
\section{Methods}{

\code{$get(extension)} Get default reader for a given extension
\code{$set(extension, f)} Set the default reader


\code{$get(class_in, extension)} Get default writer
\code{$set(class_in, extension, f)} Set the default writer
}

\keyword{datasets}
