% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{tfio}
\alias{tfio}
\alias{tfio-package}
\title{TensorFlow IO API for R}
\description{
This library provides an R interface to the
\href{https://github.com/tensorflow/io}{TensorFlow IO} API
that provides datasets and filesystem extensions maintained by SIG-IO.
}
