% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_dataset.R
\name{tiff_dataset}
\alias{tiff_dataset}
\title{Create a \code{TIFFDataset}.}
\usage{
tiff_dataset(filenames)
}
\arguments{
\item{filenames}{A \code{tf.string} tensor containing one or more filenames.}
}
\description{
A TIFF Image File Dataset that reads the TIFF file.
}
\examples{
\dontrun{
dataset <- tiff_dataset(
    filenames = list("testdata/small.tiff")) \%>\%
  dataset_repeat(1)

sess <- tf$Session()
iterator <- make_iterator_one_shot(dataset)
next_batch <- iterator_get_next(iterator)

until_out_of_range({
  batch <- sess$run(next_batch)
  print(batch)
})
}

}
