% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write}
\alias{write}
\title{Write stuff out of R}
\usage{
write(object, file, extension = tools::file_ext(file), ...,
  writer = NULL)
}
\arguments{
\item{object}{the data to be written out}

\item{extension}{the file extension (e.g. '.csv') used to find the default writer}

\item{...}{additional arguments to be passed to the writer}

\item{writer}{optional override writer to bypass default one}

\item{the}{file name for the data to be written to}
}
\description{
Write stuff out of R
}
